+++ Convert-psake.ps1
@@ -4,8 +4,7 @@
 <#
 .Synopsis
 	Converts psake build scripts to Invoke-Build.
-	Invoke-Build - Build Automation in PowerShell
-	Copyright (c) 2011-2017 Roman Kuzmin
+	Copyright (c) Roman Kuzmin
 
 .Description
 	The script converts the specified psake script to Invoke-Build script code.