+++ Show-BuildGraph.ps1
@@ -2,8 +2,7 @@
 <#
 .Synopsis
 	Shows Invoke-Build task graph using Graphviz.
-	Invoke-Build - Build Automation in PowerShell
-	Copyright (c) 2011-2017 Roman Kuzmin
+	Copyright (c) Roman Kuzmin
 
 .Description
 	Requires: