+++ Tasks/Import/.build.ps1
@@ -23,16 +23,18 @@ $MyProp1 = 'prop1'
 $env:MyProp2 = 'prop2'
 
 # Import tasks by dot-sourcing available task scripts
-foreach($_ in Get-ChildItem MyScript/*.tasks.ps1) {. $_}
+foreach($file in Get-ChildItem MyScript/*.tasks.ps1) {. $file}
 
 ### EXAMPLE 2. Import from a module with tasks
 
 # Required variable
 $MyModuleParam = 'param1'
 
-# Import the module and dot-source its tasks
+# Import the module and dot-source its provided tasks
+# In this example we can just call: . MyModule.tasks
+# But let's pretend we know just the pattern *.tasks
 Import-Module ./MyModule
-. MyModule.tasks
+foreach($file in Get-Command *.tasks -Module MyModule) {. $file}
 
 ### MAIN SCRIPT. Define own tasks