+++ Tasks/Paket/Project.build.ps1
@@ -8,14 +8,14 @@
 
 	This command invokes the task Build defined in this script.
 	The required packages are downloaded on the first call.
-	Then the task is invoked by local Invoke-Build.
+	Then Build is invoked by local Invoke-Build.
 
 .Example
 	PS> Invoke-Build Build
 
-	This also invokes the task Build defined in this script. But:
+	It also invokes the task Build defined in this script. But:
 	- It is invoked by global Invoke-Build.
-	- It does not install packages.
+	- It does not check or install packages.
 #>
 
 param(
@@ -26,12 +26,13 @@ param(
 
 # Direct call: ensure packages and call the local Invoke-Build
 
-if ($MyInvocation.ScriptName -notlike '*Invoke-Build.ps1') {
+if ([System.IO.Path]::GetFileName($MyInvocation.ScriptName) -ne 'Invoke-Build.ps1') {
 	$ErrorActionPreference = 'Stop'
 	$ib = "$PSScriptRoot/packages/Invoke-Build/tools/Invoke-Build.ps1"
 
 	# install packages
 	if (!(Test-Path -LiteralPath $ib)) {
+		'Installing packages...'
 		& $PSScriptRoot/.paket/paket.exe install
 		if ($LASTEXITCODE) {throw "paket exit code: $LASTEXITCODE"}
 	}
@@ -56,4 +57,11 @@ task Init {
 # Synopsis: Remove temporary stuff.
 task Clean {
 	Remove-Item packages, paket-files, paket.lock -Force -Recurse -ErrorAction 2
+}
+
+# import the downloaded task library and use the custom task `ask`
+. paket-files\nightroman\Invoke-Build\Tasks\Ask\Ask.tasks.ps1
+
+ask Confirm -Prompt Confirm... {
+	'Confirmed...'
 }